Move static inner classes from LuceneQueryBuilder to top level
 - Reduces the level of nesting inside the LuceneQueryBuilder.

 - Makes it clearer that those query components have a single
 responsibility and could be unit tested.
