+++ sql/src/main/java/io/crate/lucene/FunctionLiteralPair.java
@@ -1,0 +1,70 @@
+/*
+ * Licensed to Crate under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.  Crate licenses this file
+ * to you under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.  You may
+ * obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
+ *
+ * However, if you have executed another commercial license agreement
+ * with Crate these terms will supersede the license and you may use the
+ * software solely pursuant to the terms of the relevant commercial
+ * agreement.
+ */
+
+package io.crate.lucene;
+
+import io.crate.data.Input;
+import io.crate.expression.symbol.Function;
+import io.crate.expression.symbol.Symbol;
+
+class FunctionLiteralPair {
+
+    private final String functionName;
+    private final Function function;
+    private final Input input;
+
+    FunctionLiteralPair(Function outerFunction) {
+        assert outerFunction.arguments().size() == 2 : "function requires 2 arguments";
+        Symbol left = outerFunction.arguments().get(0);
+        Symbol right = outerFunction.arguments().get(1);
+
+        functionName = outerFunction.info().ident().name();
+
+        if (left instanceof Function) {
+            function = (Function) left;
+        } else if (right instanceof Function) {
+            function = (Function) right;
+        } else {
+            function = null;
+        }
+
+        if (left.symbolType().isValueSymbol()) {
+            input = (Input) left;
+        } else if (right.symbolType().isValueSymbol()) {
+            input = (Input) right;
+        } else {
+            input = null;
+        }
+    }
+
+    boolean isValid() {
+        return input != null && function != null;
+    }
+
+    Input input() {
+        return input;
+    }
+
+    String functionName() {
+        return functionName;
+    }
+}
